Renamed class files Updated use of namespaced classes
