+++ src/classes/Controllers/FrontendController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SeriouslySimplePodcasting\Frontend;
+namespace SeriouslySimplePodcasting\Controllers;
 
 // Exit if accessed directly.
 if ( ! defined( 'ABSPATH' ) ) {
@@ -15,7 +15,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  * @package     SeriouslySimplePodcasting/Classes
  * @since       1.0
  */
-class Frontend {
+class FrontendController {
 
 	// @todo reference prior to analytics launch
 	public $style_guide = array(
